requirements.txt
pandas==2.2.1
openpyxl==3.1.2
playwright==1.42.0
gspread==5.12.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0 